Delete confirmation

+++ src/main/java/seedu/command/DeleteContactCommand.java
@@ -2,12 +2,14 @@ package seedu.command;
 
 import seedu.contact.Contact;
 import seedu.parser.FailedCommandType;
+import seedu.parser.IndexParser;
 import seedu.ui.TextUi;
 import seedu.ui.ExceptionTextUi;
 import seedu.command.FailedCommand;
 
 public class DeleteContactCommand extends Command {
     private final int deletedIndex;
+    private final String confirmMessage = "";
 
     public DeleteContactCommand(int deletedIndex) {
         this.deletedIndex = deletedIndex;
@@ -17,11 +19,22 @@ public class DeleteContactCommand extends Command {
         return deletedIndex;
     }
 
-    public void execute() {
-        try {
-            Contact deletedContact = contactList.getContactAtIndex(deletedIndex);
+    private void deleteOnConfirmation(Contact deletedContact) {
+        // ask for confirmation to delete from user
+        String userDeleteConfirmation = TextUi.getUserDeleteConfirmation(deletedContact, deletedIndex);
+        if (userDeleteConfirmation.equalsIgnoreCase("y")) {
             this.contactList.deleteContact(deletedIndex);
             TextUi.deleteContactMessage(deletedContact.getName(), contactList.getListSize());
+        } else {
+            TextUi.cancelDeleteContactMessage();
+        }
+    }
+
+    public void execute() {
+        try {
+            // throws IndexOutOfBoundsException if index is outside of the range
+            Contact deletedContact = IndexParser.getContactFromIndex(deletedIndex, contactList);
+            deleteOnConfirmation(deletedContact);
         } catch (IndexOutOfBoundsException e) {
             ExceptionTextUi.numOutOfRangeMessage(contactList.getListSize());
         }
